Fedora kinda maybe stuff

+++ build.sh
@@ -1,6 +1,14 @@
 #!/bin/bash
 
-sudo apt-get install clang yasm genext2fs build-essential wget libmpfr-dev libmpc-dev libgmp-dev qemu autoconf automake texinfo
+if [ -f /etc/lsb-release ]; then
+    sudo apt-get install clang yasm genext2fs build-essential wget libmpfr-dev libmpc-dev libgmp-dev qemu autoconf automake texinfo
+elif [ -f /etc/fedora-release ]; then
+    sudo yum groupinstall 'Development Tools'
+    sudo yum groupinstall 'Development Libraries'
+    sudo yum install clang yasm mpfr-devel libmpc-devel gmp-devel
+    echo "Warning: Fedora is unsupported in this script. Be careful!"
+    echo "For best results, follow the steps in the script manually."
+fi
 # Build the toolchain:
 pushd toolchain
 ./prepare.sh
